backend/requirements.txt
Django>=4.0
djangorestframework>=3.14.0
django-celery-beat
celery>=5.0
redis>=4.0
streamlit
requests
python-dotenv
django-cors-headers
psycopg2-binary
gunicorn